+++ jsonc/deno.json
@@ -2,7 +2,6 @@
   "name": "@std/jsonc",
   "version": "0.224.0",
   "exports": {
-    ".": "./mod.ts",
-    "./parse": "./parse.ts"
+    ".": "./mod.ts"
   }
 }
